requirements.txt

Flask==2.3.3
python-dotenv==1.0.0
openai==0.28.1
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
